pyproject.toml
[project]
name = "corelia"
version = "0.1.0"
description = "AI-powered medical note processing for French healthcare"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Documentation
    "mkdocs>=1.6.1",
    "mkdocs-git-revision-date-localized-plugin>=1.4.7",
    "mkdocs-material>=9.6.19",
    
    # Core ML libraries
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
    "peft>=0.6.0",
    
    # MLflow for experiment tracking
    "mlflow>=2.8.0",
    
    # Hugging Face Hub
    "huggingface-hub>=0.17.0",
    
    # Evaluation and metrics
    "evaluate>=0.4.0",
    "scikit-learn>=1.3.0",
    
    # Utilities
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
